requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI and Resume Processing
openai==1.3.5
PyPDF2==3.0.1
pdfplumber==0.9.0
langchain==0.0.335

# Job Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
scrapy==2.11.0

# Notifications
twilio==8.10.0
sendgrid==6.10.0

# Database and Caching
redis==5.0.1
celery==5.3.4

# Utilities
pandas==2.1.3
numpy==1.25.2
schedule==1.2.0
python-crontab==3.0.0